requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.1
SQLAlchemy==1.4.32
urllib3==1.26.9
